+++ src/muibook/index.ts
@@ -108,7 +108,6 @@ import "../components/mui-form-section";
 import "../components/mui-form-section-footer";
 import "../components/mui-form-group";
 import "../components/mui-form-message";
-import "../components/mui-form-hint";
 import "../components/mui-hint";
 import "../components/mui-prompt";
 import "../components/mui-prompt-toggle";